* Move video code maps to parsers.system * Convert video converter and black_setup commands to use AVRCodeBoolMap * Convert video parsers to use code maps directly * Simplify set_video_settings as code maps now available for all video properties
